pyproject.toml
[tool.poetry]
name = "repo_standards_validator"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
structlog = "24.4.0"
pygithub = "2.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
ruff = "0.8.6"
vulture = "2.14"
zizmor = "1.0.0"
check-jsonschema = "0.30.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]

ignore = [
  "COM812",  # Ignore due to conflict with Ruff formatter
  "ISC001",  # Ignore due to conflict with Ruff formatter
  "PLR2004", # Ignore magic value
  "D104",    # Ignore missing docstring in public package
  "D100",    # Ignore missing docstring in public module
  "SIM112",  # Ignore Lowercase environment variables (used for GitHub actions)
]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"**test_*.py" = ["S101", "D101", "D102", "D103", "SLF001", "FBT001"]

[tool.ruff.lint.pydocstyle]
convention = "google" # Use Google docstring convention.

[tool.ruff.lint.isort]
known-first-party = ["tests"]

